% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataBench.R
\name{benchOOM}
\alias{benchOOM}
\title{harness for out-of-memory benchmarking}
\usage{
benchOOM(NR = 5000, NC = 100, times = 5, inseed = 1234,
  methods = list(.h5RoundTrip, .ffRoundTrip, .slRoundTrip, .dtRoundTrip,
  .bmRoundTrip))
}
\arguments{
\item{NR}{numeric number of rows of matrix to process}

\item{NC}{numeric number of columns}

\item{times}{numeric, as used by \code{\link[microbenchmark]{microbenchmark}}, number of times to execute for averaging}

\item{inseed}{numeric, as used by \code{\link[base]{set.seed}}}

\item{methods}{a list of functions, each having arguments \code{x} and \code{intimes}, with \code{x}
the matrix being processed and \code{intimes} to be passed to microbenchmark for \code{times}}
}
\description{
harness for out-of-memory benchmarking
}
\note{
the \code{methodList} function collects some example 'round trip' simulation functions
}
\examples{
methodList()$bigm
names(methodList())
benchOOM(methods=methodList())
}
