% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportDSE.R
\docType{class}
\name{DelayedRangedSummarizedExperiment-class}
\alias{DelayedRangedSummarizedExperiment-class}
\title{marker class for delayed assay content}
\description{
marker class for delayed assay content
}
