% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bano.R
\name{get_bano_hsds}
\alias{get_bano_hsds}
\title{get an HSDS instance of banovich 450k data}
\usage{
get_bano_hsds()
}
\value{
restful SummarizedExperiment
}
\description{
get an HSDS instance of banovich 450k data
}
