% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bano.R
\name{get_bano_inmem}
\alias{get_bano_inmem}
\title{get an in-memory SummarizedExperiment}
\usage{
get_bano_inmem()
}
\value{
SummarizedExperiment instance
}
\description{
get an in-memory SummarizedExperiment
}
