% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadlit.R
\name{getLitGeu}
\alias{getLitGeu}
\title{obtain example of HDF5-backed RangedSummarizedExperiment}
\usage{
getLitGeu()
}
\description{
obtain example of HDF5-backed RangedSummarizedExperiment
}
\examples{
require("SummarizedExperiment")
litgeu = getLitGeu()
assay(litgeu[1:5,1:5])

}
