% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gobr.R
\name{gobrowse}
\alias{gobrowse}
\title{Use a shiny gadget to query user on a GO category and return entrez ids}
\usage{
gobrowse(src_o, ont = "BP", BPini = "A-L", iniGO = "abscission")
}
\arguments{
\item{src_o}{instance of src_organism in Organism.dplyr}
}
\description{
Use a shiny gadget to query user on a GO category and return entrez ids
}
