% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5serv.R
\name{h5serv}
\alias{h5serv}
\title{prototypical query to HDF5 server established by Landmark Bioinfo (Shweta Gopaulakrishnan, reshg@channing.harvard.edu, Vince Carey, stvjc@channing.harvard.edu)}
\usage{
h5serv(dsn = "assay001", rows = 1:3, cols = 1:5,
  h5id = "assays.hdfgroup.org", endpoint = "http://54.163.220.201:5000")
}
\arguments{
\item{dsn}{internal HDF5 name of dataset}

\item{rows}{vector of row selection, only min and max are used to define a range, all data in range retrieved}

\item{cols}{vector of column selection, only min and max are used to define a range, all data in range retrieved}

\item{h5id}{name of HDF5 resource as identified by server}

\item{endpoint}{URL for server}
}
\description{
There is no guarantee of availability of this service.
}
\note{
You will need python h5pyd support installed, see \url{https://github.com/HDFGroup/h5pyd}
}
\examples{
h5serv()
}
