% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataBench.R
\name{methodList}
\alias{methodList}
\title{helper for creating a methodlist}
\usage{
methodList(methods = c("hdf5", "bigm", "sqlite", "ff"))
}
\arguments{
\item{methods}{a character vector with tags for available round trip methods}
}
\description{
helper for creating a methodlist
}
