% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportmicro_H5bano.R
\name{micro_H5bano}
\alias{micro_H5bano}
\title{microbenchmark retrieval of banovich assay data from HDF Server. [running on ec2 instance and schion]}
\usage{
micro_H5bano(serverURL = serverURL, host = host, times = times, i = i,
  j = j)
}
\arguments{
\item{serverURL}{character including the port}

\item{host}{domain name of the data file on the server}

\item{times}{benchmarking time nevals}

\item{i}{character corresponding to row [e.g: "0:3"]}

\item{j}{character correspoding to coloumn [e,g: "0:3"]}
}
\value{
a m1 with the microbenchmarking results [timings]
}
\examples{
 
\dontrun{
   serverURL = "http://54.163.220.201:5000"
   host = "assays.hdfgroup.org"
   micro_H5bano(serverURL = serverURL,host = host,times=5, i="0:3",j="0:3")
}
}
\author{
Shweta Gopaulakrishnan <reshg@channing.harvard.edu>
}
