% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setContexts.R
\name{resolve}
\alias{resolve}
\title{translate a token using annotation context}
\usage{
resolve(sym)
}
\arguments{
\item{sym}{a symbol to be resolved}
}
\description{
translate a token using annotation context
}
