% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportDSE.R
\name{rhconvert}
\alias{rhconvert}
\title{dump assay of RangedSummarizedExperiment to Delayed HDF5 and return suitable instance}
\usage{
rhconvert(rse, file, name, ...)
}
\arguments{
\item{rse}{RangedSummarizedExperiment instance}

\item{file}{character string with path to HDF5 store on disk}

\item{name}{character string with name of HDF5 dataset}
}
\description{
dump assay of RangedSummarizedExperiment to Delayed HDF5 and return suitable instance
}
\examples{
\dontrun{
 library(geuvPack)
 data(geuFPKM)
 delgeu = rhconvert( geuFPKM, "geuFPKM.h5", "geuFPKM" )
 delgeu
}

}
