% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportBTSE.R
\name{writeBigTable}
\alias{writeBigTable}
\title{propagate information from a RangedSummarizedExperiment to google BigTable}
\usage{
writeBigTable(x, src, featureKeyName, sampleKeyName, tablename, ...)
}
\arguments{
\item{x}{RangedSummarizedExperiment}

\item{src}{instance of src_bigrquery}

\item{featureKeyName}{character string that is used for joining features to metadata}

\item{sampleKeyName}{character string that is used for joining samples to feature data}

\item{tablename}{character string used as a prefix, suffixes _assay, _ranges, _coldata will be appended in the back end}

\item{\dots}{not used}
}
\description{
propagate information from a RangedSummarizedExperiment to google BigTable
}
\examples{
\dontrun{
 require(geuvPack)
 data(geuFPKM) # following will generate charges,
 # assumes that 'bq' is an authenticated 'src_bigquery' instance
 bq = writeBigTable( geuFPKM, bq, "GENEID", "SAMPLE", "geuFPKM" )
 bq
}

}
