% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportDSE.R
\name{writeHDF5Dataset}
\alias{writeHDF5Dataset}
\alias{"writeHDF5Dataset-RangedSummarizedExperiment"}
\title{support for HDF5-backed SummarizedExperiment}
\usage{
writeHDF5Dataset(x, file, name, ...)
}
\arguments{
\item{x}{\code{\link[SummarizedExperiment]{SummarizedExperiment-class}} instance}

\item{file}{character string identifying target file to hold HDF5}

\item{name}{character string identifying dataset}

\item{\dots}{passed down}
}
\description{
support for HDF5-backed SummarizedExperiment
}
