% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportOracleSE.R
\name{writeOracleTable}
\alias{writeOracleTable}
\title{propagate information from a RangedSummarizedExperiment to Oracle}
\usage{
writeOracleTable(x, con, featureKeyName, sampleKeyName, tablename, ...)
}
\arguments{
\item{x}{RangedSummarizedExperiment}

\item{con}{connection to oracle database}

\item{featureKeyName}{character string that is used for joining features to metadata}

\item{sampleKeyName}{character string that is used for joining samples to feature data}

\item{tablename}{character string used as a prefix, suffixes _assay, _ranges, _coldata will be appended in the back end}

\item{\dots}{not used}
}
\description{
propagate information from a RangedSummarizedExperiment to Oracle
}
\examples{
\dontrun{
 require(geuvPack)
 data(geuFPKM) # following will generate charges,
 # assumes that connection has been made to the database
 writeOracleTable( geuFPKM, con, "GENEID", "SAMPLE", "geuFPKM" )
}

}
