% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setContexts.R
\name{setContexts}
\alias{setContexts}
\title{establish an (ordered) list of hashes or functions that translate symbols}
\usage{
setContexts(conlist)
}
\arguments{
\item{conlist}{a list of functions or named vectors; no validity checking}
}
\description{
establish an (ordered) list of hashes or functions that translate symbols
}
